% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{generate_col_types}
\alias{generate_col_types}
\title{Assit function for readr \code{col_types} parameters}
\usage{
generate_col_types(schema)
}
\arguments{
\item{schema}{pre-specified schema}
}
\value{
list of col types
}
\description{
Assit function for readr \code{col_types} parameters
}
