% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate input data with the provided schema}
\usage{
validate(data, schema, quiet = FALSE)
}
\arguments{
\item{data}{data to be validated}

\item{schema}{pre-specified schema}

\item{quiet}{TRUE when validate without console output, only return the invalid stack table}
}
\value{
a list with success indicator and stack of error messages
}
\description{
Validate input data with the provided schema
}
