% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_col}
\alias{validate_col}
\title{Validate certain cols of the input data with the provided schema}
\usage{
validate_col(data, ..., schema, quiet = FALSE)
}
\arguments{
\item{data}{data to be validated}

\item{...}{columns to be validated}

\item{schema}{pre-specified schema}

\item{quiet}{TRUE when validate without console output, only return the invalid stack table}
}
\description{
Validate certain cols of the input data with the provided schema
}
