% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_col_structure}
\alias{validate_col_structure}
\title{Validate column structures}
\usage{
validate_col_structure(data, schema, quiet = FALSE)
}
\arguments{
\item{data}{data to be validated}

\item{schema}{pre-specified schema}

\item{quiet}{TRUE when validate without console output, only return the invalid stack table}
}
\description{
Validate column structures
}
