% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_object}
\alias{validate_object}
\title{Validate the lowest object level}
\usage{
validate_object(object, object_schema)
}
\arguments{
\item{object}{the cell object in the incoming data}

\item{object_schema}{the prespecified schema for this object}
}
\value{
a description if the object is invalid
}
\description{
Validate the lowest object level
}
