% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_obs}
\alias{validate_obs}
\title{Validate each observation}
\usage{
validate_obs(data, schema)
}
\arguments{
\item{data}{data to be validated}

\item{schema}{pre-specified schema}
}
\description{
Validate each observation
}
